// src/enums/flex.enum.ts
export enum AlignItems {
    FlexStart = 'flex-start',
    Center = 'center',
    FlexEnd = 'flex-end',
    Stretch = 'stretch',
};

export enum JustifyContent {
    FlexStart = 'flex-start',
    Center = 'center',
    FlexEnd = 'flex-end',
    SpaceBetween = 'space-between',
    SpaceAround = 'space-around',
};

export enum FlexDirection {
    Row = 'row',
    Column = 'column',
    RowReverse = 'row-reverse',
    ColumnReverse = 'column-reverse',
}